IS-386 Merged newer version of gee_test into src/fusion/tools
