--- earth_enterprise/qa/gee_test/add_portable_tester.sh
@@ -1,19 +1,0 @@
-#!/bin/bash
-#
-# Copyright 2017 Google Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-# TODO: Consider renaming test_portable.py gold_test.py.
-cp ../../portableglobe/servers/test/test_portable.py tests/gold_tests/gold_test.py
