--- earth_enterprise/qa/gee_test/tutorial_files/providers.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no" ?>
-<gstProviderSet>
-
-  <items>
-    <item>
-      <name>USGS Imagery</name>
-      <key>USGS-I</key>
-      <copyright>Imagery © 2005 USGS</copyright>
-      <copyrightVerticalPos>-1</copyrightVerticalPos>
-      <report>0</report>
-      <copyrightPriority>256</copyrightPriority>
-      <id>1</id>
-    </item>
-    <item>
-      <name>I3</name>
-      <key>I3</key>
-      <copyright>Imagery © 2005 i3</copyright>
-      <copyrightVerticalPos>-1</copyrightVerticalPos>
-      <report>0</report>
-      <copyrightPriority>256</copyrightPriority>
-      <id>2</id>
-    </item>
-    <item>
-      <name>USGS Maps</name>
-      <key>USGS-M</key>
-      <copyright>Roads © 2005 USGS</copyright>
-      <copyrightVerticalPos>-1</copyrightVerticalPos>
-      <report>0</report>
-      <copyrightPriority>256</copyrightPriority>
-      <id>3</id>
-    </item>
-    <item>
-      <name>GNIS/US Census Bureau</name>
-      <key>GNIS</key>
-      <copyright>Population © 2005 GNIS</copyright>
-      <copyrightVerticalPos>-1</copyrightVerticalPos>
-      <report>0</report>
-      <copyrightPriority>256</copyrightPriority>
-      <id>4</id>
-    </item>
-    <item>
-      <name>USGS POIs</name>
-      <key>USGS-P</key>
-      <copyright>POIs © 2004 USGS</copyright>
-      <copyrightVerticalPos>-1</copyrightVerticalPos>
-      <report>0</report>
-      <copyrightPriority>256</copyrightPriority>
-      <id>5</id>
-    </item>
-    <item>
-      <name>USGS Terrain</name>
-      <key>USGS-T</key>
-      <copyright>Terrain © USGS 2003</copyright>
-      <copyrightVerticalPos>-1</copyrightVerticalPos>
-      <report>0</report>
-      <copyrightPriority>256</copyrightPriority>
-      <id>6</id>
-    </item>
-    <item>
-      <name>NASA Imagery</name>
-      <key>NASA</key>
-      <copyright>Imagery courtesy NASA's Earth Observatory</copyright>
-      <copyrightVerticalPos>-1</copyrightVerticalPos>
-      <report>0</report>
-      <copyrightPriority>256</copyrightPriority>
-      <id>7</id>
-    </item>
-  </items>
-
-  <next_id>8</next_id>
-
-</gstProviderSet>
