+++ earth_enterprise/src/fusion/tools/gee_test/README
@@ -1,17 +1,17 @@
-#
-# Copyright 2017 Google Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+// Copyright 2017 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 
 Tool for installing Fusion and Earth Server, building tutorial databases (2d, 3d, and 3d Time Machine), and running tests against these databases.
 
@@ -58,22 +58,22 @@ Note: gee_test.py has not been updated to work with the latest installer, so the
 
 To do a basic install and test:
 
-./gee_test.py <path-to-installer-tarball>
+sudo ./gee_test.py <path-to-installer-tarball>
 
 or
 
-./gee_test.py <path-to-installer-directory>
+sudo ./gee_test.py <path-to-installer-directory>
 
 
 To install without building tutorial dbs:
 
-./gee_test.py --notutorial --nopublish --notests <path-to-installer-tarball>
+sudo ./gee_test.py --notutorial --nopublish --notests <path-to-installer-tarball>
 
 
 To do a clean install with tests:
 
-./gee_test.py --clean --uninstall <path-to-installer-directory>
+sudo ./gee_test.py --clean --uninstall <path-to-installer-directory>
 
 To run only tests:
 
-./gee_test.py --noserver --nofusion --notutorial --nopublish <path-to-installer-directory>
+sudo ./gee_test.py --noserver --nofusion --notutorial --nopublish <path-to-installer-directory>
