+++ earth_enterprise/src/fusion/tools/gee_test/gee_test.py
@@ -13,16 +13,20 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+#
 
 """Install Fusion and Earth Server, build tutorial databases, and test."""
 
 import getpass
 import os
+import socket
 import sys
 import time
 
 from tests.common import utils
 
+
+HOSTNAME = socket.getfqdn()
 FUSION_SILENT_INSTALL_TEMPLATE = "installfusion.properties"
 SERVER_SILENT_INSTALL_TEMPLATE = "installserver.properties"
 TMP_DIR = "/tmp"
@@ -32,9 +36,17 @@ WAIT_FOR_SERVER_STOP = 5
 WAIT_FOR_SERVER_START = 5
 MIN_DISK_SPACE = 2000  # ~2GB
 
+TUTORIAL_DBS = ("SFDb_3d", "SFDb_3d_TM", "SF_2d_Merc")
+
+GEVOL_DIR = "/usr/local/google/gevol_test"
+PUBLISHED_DBS_STREAM_SPACE_DIR = os.path.join(
+    GEVOL_DIR,
+    "published_dbs/stream_space")
+
 
 def CheckEnvironment(no_tests):
   """Make sure that we are not root and that we have prodaccess."""
+  print "Hostname: ", HOSTNAME
   if utils.DiskSpace("/tmp") < MIN_DISK_SPACE:
     print "Insufficient disk space."
     exit(0)
@@ -42,6 +54,11 @@ def CheckEnvironment(no_tests):
     print "Please do not run as root."
     print "In order to run tests, you must use your own prodaccess."
     exit(0)
+  if not no_tests:
+    prodcertstatus = utils.ExecuteCmd("prodcertstatus")
+    print prodcertstatus
+    if "No valid" in prodcertstatus:
+      utils.ExecuteCmd("prodaccess --nog")
 
 
 def StopFusion():
@@ -101,8 +118,9 @@ def Clean(force=False):
     if answer != "y":
       print "Skipping clean ..."
       return
-  utils.ExecuteCmd("sudo rm -rf /usr/local/google/gevol_test/assets/*")
-  utils.ExecuteCmd("sudo rm -rf /usr/local/google/gevol_test/published_dbs/*")
+
+  gevol_dir = GEVOL_DIR
+  utils.ExecuteCmd("sudo rm -rf %s/*" % gevol_dir)
   utils.ExecuteCmd("sudo rm -rf /opt/google/gehttpd")
   utils.ExecuteCmd("sudo rm -rf /var/opt/google/pgsql")
   utils.ExecuteCmd("sudo rm -rf /var/opt/google/fusion-backups")
@@ -156,9 +174,12 @@ def BuildTutorial(base_dir):
   utils.ExecuteCmd("env PATH=\"$PATH:/opt/google/bin\" ./run_tutorial.sh")
 
 
-def WaitForTutorialBuilds():
-  """Wait until Tutorial databases are built."""
-  dbs = ("SFDb_3d", "SFDb_3d_TM", "SF_2d_Merc")
+def WaitForTutorialBuilds(dbs):
+  """Wait until Tutorial databases are built.
+
+  Args:
+    dbs: a list of database names.
+  """
   succeeded = {}
   for db in dbs:
     succeeded[db] = False
@@ -179,39 +200,146 @@ def WaitForTutorialBuilds():
       time.sleep(BUILD_CHECK_WAIT_IN_SECS)
 
 
-def PushAndPublishTutorialDbs():
-  """Publish the tutorial databases."""
-  dbs = ("SFDb_3d", "SFDb_3d_TM", "SF_2d_Merc")
+def GetGedbPath(db_name):
+  """Gets gedb/mapdb path for specified database.
+
+  Args:
+    db_name: a database name.
+
+  Returns:
+    a gedb/mapdb path.
+  """
+  result = utils.ExecuteCmd(
+      "/opt/google/bin/gequery Databases/%s --outfiles" % db_name)
+  db_path = ""
+  for line in result.split("\n"):
+    if line.endswith("/gedb") or line.endswith("/mapdb"):
+      db_path = line
+      break
+
+  return db_path
+
+
+def GetGedbPathForDbs(dbs):
+  """Gets gedb/mapdb path for list of databases.
+
+  Args:
+    dbs: a list of database names.
+
+  Returns:
+    a list of gedb/mapdb paths.
+  """
+  gedb_paths = []
+  for db_name in dbs:
+    gedb_paths.append(GetGedbPath(db_name))
+
+  return gedb_paths
+
+
+def GetPublishGedbPathForDbs(dbs):
+  """Gets publish gedb/mapdb path for list of databases.
+
+  Args:
+    dbs: a list of database names/paths.
+
+  Returns:
+    a list of publish gedb/mapdb paths.
+  """
+  publish_gedb_paths = []
+  publish_dbs_base_path = os.path.join(PUBLISHED_DBS_STREAM_SPACE_DIR, HOSTNAME)
+  for db_name in dbs:
+    if db_name and os.path.isabs(db_name):
+      gedb_path = db_name
+    else:
+      gedb_path = GetGedbPath(db_name)
+
+    publish_gedb_paths.append("{}{}".format(publish_dbs_base_path, gedb_path))
+
+  return publish_gedb_paths
+
+
+def PushAndPublishDbs(dbs):
+  """Push and publish list of databases.
+
+  Args:
+    dbs: list of databases.
+
+  Raises:
+    Exception: when unable to find database.
+  """
   # TODO: Be smarter about things that are already pushed and
   #  published. Right now, publishing in this way with unpublish if there is
   #  something already using the target.
   for db in dbs:
-    result = utils.ExecuteCmd(
-        "sudo /opt/google/bin/gequery Databases/%s --outfiles" % db)
-    db_path = ""
-    for line in result.split("\n"):
-      if line.endswith("/gedb"):
-        db_path = line
-        break
-      elif line.endswith("/mapdb"):
-        db_path = line
-        break
+    db_path = GetGedbPath(db)
     if not db_path:
       raise Exception("Unable to find built database %s." % db)
-    print "push and publish", db_path
-    utils.ExecuteCmd("sudo /opt/google/bin/geserveradmin --adddb %s" % db_path)
-    utils.ExecuteCmd("sudo /opt/google/bin/geserveradmin --pushdb %s" % db_path)
+
+    print "push and publish: ", db_path
+    utils.ExecuteCmd("/opt/google/bin/geserveradmin --adddb %s" % db_path)
+    utils.ExecuteCmd("/opt/google/bin/geserveradmin --pushdb %s" % db_path)
     # TODO: Use web api.
     utils.ExecuteCmd(
-        "sudo /opt/google/bin/geserveradmin --publishdb %s --targetpath %s" %
+        "/opt/google/bin/geserveradmin --publishdb %s --targetpath %s" %
         (db_path, db))
 
 
-def RunTests(name):
-  """Run all .py files in the current directory as tests."""
+def RunGeindexcheck(db_name):
+  """Runs geindexcheck for specified database.
+
+  Args:
+    db_name: database name/path.
+  """
+
+  if db_name and os.path.isabs(db_name):
+    db_path = os.path.normpath(db_name)
+  else:
+    db_path = os.path.join("Databases", db_name)
+
+  result = utils.ExecuteCmd(
+      "/opt/google/bin/geindexcheck --mode all --database %s" % db_path,
+      do_log=False, err2out=True)
+
+  if "Fusion Fatal:" in result:
+    print "geindexcheck FAILED for database: ", db_path
+    print "Result: ", result
+  else:
+    print "SUCCESS"
+
+
+def RunGeindexcheckForDbs(dbs):
+  """Runs geindexcheck for list of databases.
+
+  Args:
+    dbs: list of database names/paths.
+  """
+  for db_name in dbs:
+    RunGeindexcheck(db_name)
+
+
+def GetTests(base_dir):
+  """Get tests from cns and unpack them."""
+  utils.ChDir(base_dir)
+  test_sets = [
+      "/cns/ih-d/home/bpeterson/gee_tests/gold_tests.tgz"
+      ]
+  for test_set in test_sets:
+    utils.ExecuteCmd("fileutil cp %s test_set.tgz" % test_set)
+    utils.ExecuteCmd("tar -xzf test_set.tgz")
+    for path in os.listdir("test_set"):
+      utils.ExecuteCmd("cp -r test_set/%s tests/" % path)
+    utils.ExecuteCmd("rm test_set.tgz")
+    utils.ExecuteCmd("rm -rf test_set")
+    print "installed test set %s" % test_set
+
+
+def RunTests(base_dir):
+  """Runs tests in the test directory."""
+  utils.ChDir(base_dir)
   num_tests = 0
   successes = 0
-  for test in os.listdir("."):
+  for test in os.listdir("tests"):
+    utils.ChDir("%s/tests" % base_dir)
     if not test.startswith("__") and test.endswith(".py"):
       num_tests += 1
       result = utils.ExecuteCmd("./%s" % test)
@@ -223,32 +351,8 @@ def RunTests(name):
             successes += 1
           print result
           break
-  print "***** Testing: %s *****" % name
-  print "%d of %d tests passed.\n" % (successes, num_tests)
-
-
-def RunTestSets(base_dir):
-  """Extract and run test sets."""
-  print "** For default test sets, please run:"
-  print "fileutil cp /cns/ih-d/home/gee/test_sets/*.tgz test_sets/"
-  utils.ChDir("%s/test_sets" % base_dir)
-  for test_set_tarball in os.listdir("."):
-    if test_set_tarball.endswith(".tgz"):
-      # Extract tests
-      utils.ExecuteCmd("tar -xzf %s" % test_set_tarball)
-      utils.ChDir("test_set")
-      # Run tests
-      RunTests(test_set_tarball)
-      utils.ChDir("..")
-      # Clean up
-      utils.ExecuteCmd("rm -rf test_set")
-
-
-def RunLocalTests(base_dir):
-  """Runs tests in the test directory."""
-  utils.ChDir(base_dir)
-  utils.ChDir("%s/tests" % base_dir)
-  RunTests("local")
+  print "*****"
+  print "%d of %d tests passed." % (successes, num_tests)
 
 
 def Flag(flag, argv):
@@ -269,9 +373,12 @@ def Usage(cmd):
   print "       nofusion   (optional) - skip Fusion install."
   print "       noserver   (optional) - skip Earth Server install."
   print "       notutorial (optional) - skip building tutorial databases."
+  print ("       nogeindexcheck (optional) - skip geindexcheck running for "
+         "tutorial databases.")
   print ("       nopublish  (optional) - skip pushing and publishing tutorial "
          "databases.")
   print "       notests    (optional) - skip running tests."
+  print "       nogettests (optional) - skip getting tests from cns."
   print "       force      (optional) - do not ask for confirmation of clean."
   exit(0)
 
@@ -331,23 +438,39 @@ def main(argv):
   StartServer()
   time.sleep(WAIT_FOR_SERVER_START)
 
+  dbs = TUTORIAL_DBS
+
   # Build tutorial.
   if not Flag("notutorial", argv):
     BuildTutorial(utils.BaseDir())
   else:
     print "Skipping building tutorial databases."
 
+  if not Flag("nogeindexcheck", argv):
+    WaitForTutorialBuilds(dbs)
+    RunGeindexcheckForDbs(dbs)
+  else:
+    print "Skipping geindexcheck for tutorial databases"
+
   # Wait for build, then push and publish tutorial dbs.
   if not Flag("nopublish", argv):
-    WaitForTutorialBuilds()
-    PushAndPublishTutorialDbs()
+    WaitForTutorialBuilds(dbs)
+    PushAndPublishDbs(dbs)
+
+    print "Run geindexcheck for pushed/published DBs:"
+    publish_gedb_paths = GetPublishGedbPathForDbs(dbs)
+    RunGeindexcheckForDbs(publish_gedb_paths)
   else:
     print "Skipping pushing and publishing of tutorial databases."
 
   # Run tests.
   if not Flag("notests", argv):
-    RunLocalTests(utils.BaseDir())
-    RunTestSets(utils.BaseDir())
+    if not Flag("nogettests", argv):
+      GetTests(utils.BaseDir())
+    else:
+      print "Skipping updating of tests from cns."
+
+    RunTests(utils.BaseDir())
   else:
     print "Skipping tests."
