+++ earth_enterprise/src/fusion/tools/gee_test/tests/README
@@ -1,19 +1,18 @@
-#
-# Copyright 2017 Google Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
+// Copyright 2017 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
 
 All python files in this directory will be run as tests. Each test can assume that the tutorial databases have been built and published. A test should output "SUCCESS" as its final line if it passes. All tests at this level should run without command line arguments. Tests that require command line arguments should be nested into directories here and then called with the appropriate arguments from a test here. If a test fails, it should be easy for the tester to re-run the test at lower and lower levels to help them isolate the problem.
 
-This directory is meant for exploring new tests. Well-established tests should be put into tarballs and moved into the test_sets directory. These should also be stored in cns.
+Tests will be loaded from cns base on the tests_sets list in GetTests in gee_test.py. Test sets should be world readable tarballs that extract to the directory test_set. All items in the extracted test_set directory will be moved to tests, so any files at the top level ending in .py will be executed as tests. Currently, we get all test sets before we execute them, so some care must be taken to protect file name collisions.
